+++ src/main/java/com/application/close/exception/ModbusOperationException.java
@@ -1,0 +1,20 @@
+package com.application.close.exception;
+
+import lombok.Getter;
+
+@Getter
+public class ModbusOperationException extends RuntimeException {
+
+	/**
+	 * 
+	 */
+	private static final long serialVersionUID = -825682828962349425L;
+
+	private final String message;
+
+	public ModbusOperationException(String message) {
+		super();
+		this.message = message;
+	}
+
+}