+++ src/main/java/com/application/close/links/service/BridgeExecutorServiceImpl.java
@@ -1,11 +1,13 @@
 package com.application.close.links.service;
 
+import java.time.LocalDateTime;
 import java.util.List;
 
 import org.springframework.stereotype.Service;
 
 import com.application.close.exception.BadRequestException;
 import com.application.close.exception.ResourceNotFoundException;
+import com.application.close.links.ModbusFunctionType;
 import com.application.close.links.entity.BridgeExecutor;
 import com.application.close.links.payload.BridgeExecutorPayload;
 import com.application.close.links.repo.BridgeExecutorRepo;
@@ -38,14 +40,30 @@ public class BridgeExecutorServiceImpl implements BridgeExecutorService {
 		if (executorRepo.existsByPublishTopic(executerPayload.getPublishTopic()))
 			throw new BadRequestException("Given publish topics already register for tcpDataid: " + tcpDataId);
 
+		if (!isValidFunction(executerPayload.getFunctionType()))
+			throw new BadRequestException("Given Function type invalid for tcpDataid: " + tcpDataId);
+
 		BridgeExecutor bridgeExecutor = new BridgeExecutor();
 		bridgeExecutor.setBridgeName(executerPayload.getBridgeName());
-		bridgeExecutor.setFunctionType(executerPayload.getFunctionType());
+		bridgeExecutor.setFunctionType(ModbusFunctionType.valueOf(executerPayload.getFunctionType()));
 		bridgeExecutor.setOffset(executerPayload.getOffset());
 		bridgeExecutor.setPublishTopic(executerPayload.getPublishTopic());
 		bridgeExecutor.setQuantity(executerPayload.getQuantity());
 		bridgeExecutor.setSlaveId(executerPayload.getSlaveId());
 		return executorRepo.save(bridgeExecutor);
+	}
+
+	private boolean isValidFunction(String input) {
+		if (input == null)
+			return false;
+
+		switch (input.toUpperCase()) {
+		case "READ_COILS", "READ_DISCRETE_INPUTS", "READ_HOLDING_REGISTERS", "READ_INPUT_REGISTERS":
+			return true;
+		default:
+			return false;
+		}
+
 	}
 
 	@Override
@@ -67,14 +85,24 @@ public class BridgeExecutorServiceImpl implements BridgeExecutorService {
 		if (executorRepo.existsByPublishTopic(executerPayload.getPublishTopic()))
 			throw new BadRequestException("Given publish topics already register for tcpDataid: " + tcpDataId);
 
+		if (!isValidFunction(executerPayload.getFunctionType()))
+			throw new BadRequestException("Given Function type invalid for tcpDataid: " + tcpDataId);
+
 		bridgeExecutor.setBridgeName(executerPayload.getBridgeName());
-		bridgeExecutor.setFunctionType(executerPayload.getFunctionType());
+		bridgeExecutor.setFunctionType(ModbusFunctionType.valueOf(executerPayload.getFunctionType()));
 		bridgeExecutor.setOffset(executerPayload.getOffset());
 		bridgeExecutor.setPublishTopic(executerPayload.getPublishTopic());
 		bridgeExecutor.setQuantity(executerPayload.getQuantity());
 		bridgeExecutor.setSlaveId(executerPayload.getSlaveId());
 		return executorRepo.save(bridgeExecutor);
 
+	}
+
+	@Override
+	public void updateTimesatamp(int executerId) {
+		BridgeExecutor bridgeExecutor = getById(executerId);
+		bridgeExecutor.setLastExecuted(LocalDateTime.now());
+		executorRepo.save(bridgeExecutor);
 	}
 
 	@Override