+++ src/main/java/com/application/close/mqtt/service/MqttService.java
@@ -10,7 +10,4 @@ public interface MqttService {
 
 	void disconnect(int mqttParamId);
 
-//	MqttConnectResp connectWithTLS(int mqttParamId); // TLS (SSL) connection
-//
-//	MqttConnectResp connectWithTLSAndAuth(int mqttParamId); // TLS + username/password
 }